Removed source.base.SourceGroup
